feat: trait bound-specific sync builders
