+++ rust/hyperlane-core/src/traits/indexer.rs
@@ -36,6 +36,9 @@ pub trait Indexer<T: Sized>: Send + Sync + Debug {
 }
 
 /// Interface for indexing data in sequence.
+/// SequenceAwareIndexer is an umbrella trait for all indexers types (sequence-aware and rate-limited).
+/// The rate-limited indexer doesn't need `SequenceAwareIndexer`, so impls of `SequenceAwareIndexer` just return nullish values.
+/// TODO: Refactor such that indexers aren't required to implement `SequenceAwareIndexer`
 #[async_trait]
 #[auto_impl(&, Box, Arc)]
 pub trait SequenceAwareIndexer<T>: Indexer<T> {
